// src/shared/proto/purchase.ts
export interface Purchase {
    /**
     * @generated from protobuf field: string purchases_id = 1;
     */
    purchasesId: string;
    /**
     * @generated from protobuf field: string tenants_id = 2;
     */
    tenantsId: string;
    /**
     * @generated from protobuf field: string purchase_number = 3;
     */
    purchaseNumber: string;
    /**
     * @generated from protobuf field: string status = 4;
     */
    status: string;
    /**
     * @generated from protobuf field: string users_id = 5;
     */
    usersId: string;
    /**
     * @generated from protobuf field: string events_id = 6;
     */
    eventsId: string;
    /**
     * @generated from protobuf field: int32 subtotal_cents = 7;
     */
    subtotalCents: number;
    /**
     * @generated from protobuf field: int32 fee_cents = 8;
     */
    feeCents: number;
    /**
     * @generated from protobuf field: int32 total_cents = 9;
     */
    totalCents: number;
    /**
     * @generated from protobuf field: int32 seats_reserved = 10;
     */
    seatsReserved: number;
}
// ...
export const Purchase = new Purchase$Type();